feat: Placeholder UI 추가
